Organize vite files
