--- index.html
@@ -1,19 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="UTF-8" />
-    <title>chaiNNer</title>
-
-  </head>
-  <body>
-    <div id="root">
-
-    </div>
-    <script>
-      if (global === undefined) {
-        var global = window;
-      }
-    </script>
-    <script type="module" src="/src/renderer/renderer.ts"></script>
-  </body>
-</html>
